adjusted text sizes and a bit of spacing
