+++ src/popup/components/Settings/index.tsx
@@ -1,24 +1,17 @@
 import React from "react";
 import Stack from "@mui/material/Stack";
-import Card from "../Card/Card";
 import ResetStorageSetting from "./ResetStorageSetting";
-import CustomUISettings from "./CustomUISettings";
+import HeaderClickSetting from "./HeaderClickSetting";
 import AccessTokenSetting from "./AccessTokenSetting";
 import AnimatedExpandSettingCard from "./AnimatedExpandSettingCard";
 
 export default function Settings() {
   return (
     <Stack width="100%" padding={1} spacing={1} bgcolor="whitesmoke">
-      <Card sx={{ bgcolor: "white" }}>
-        <AccessTokenSetting />
-      </Card>
-      <Card sx={{ bgcolor: "white" }}>
-        <CustomUISettings />
-      </Card>
+      <AccessTokenSetting />
+      <HeaderClickSetting />
       <AnimatedExpandSettingCard />
-      <Card sx={{ bgcolor: "white" }}>
-        <ResetStorageSetting />
-      </Card>
+      <ResetStorageSetting />
     </Stack>
   );
 }
